pyproject.toml
[project]
name = "slurp"
version = "0.1.0"
description = "Confluence RAG Dataset Generator - A distributed system for scraping Confluence pages and generating Question-Answer pairs for RAG evaluation datasets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atlassian-python-api>=3.41.0",
    "aiokafka>=0.11.0",
    "orjson>=3.10.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.1",
    "selectolax>=0.3.0",
    "nltk>=3.8.0",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.14",
    "fastapi>=0.109.0",
    "logfire>=0.28.0",
    "starlette>=0.36.0",
    "httpx>=0.26.0",
    "requests>=2.31.0",
    "aiosqlite>=0.21.0",
]

[project.optional-dependencies]
scripts = [
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
